// extensions/pokedex/src/types.ts
export interface TypeOption {
  label: string;
  value: string;
  emoji?: string;
}